requirements.txt
streamlit
langchain
langchain_groq
sentence-transformers
faiss-cpu
cohere
unstructured[all-docs]
wikipedia
PyPDF2
python-dotenv
requests 